Average mode working
I had to change the shif_corners operation to operate on floats because
otherwise it is not possible to make an average (on floats) on a shifted
image. The image has to be first shifted (so in float) and then the
average calculation can be done.

Results are stored in a vector of float in the pipeline but this solution
is temporary.