chg: added SquareInputMode and new get_X_frame_descriptor in ithread_input.hh
IThreadInput now has a field of type enum SquareInputMode which
describes if the input should be left as it is, embedded into a bigger
square or cropped into a smaller square.

There's now a distinction between get_input_frame_descriptor and
get_queue_frame_descriptor. The first one being the frame descriptor of
the original data input. The second one being the frame descriptor of
the resulting queue after taking into account the SquareInputMode.

Typically,
- SquareInputMode::NO_MODIFICATION would give the same FD.

- SquareInputMode::EMBED_INTO_SQUARE would give a square FD which side is
the biggest between the two lengths of the input FD.

- SquareInputMode::CROP_INTO_SQUARE would give a square FD which side is
the smallest between the two lengths of the input FD.